Condense login and register routes into account route
