+++ app.js
@@ -12,9 +12,7 @@ var viewEngine = require('mustache-express');
 // require routers
 var indexRouter = require('./routes/index');
 var apiRouter = require('./routes/api');
-var loginRouter = require('./routes/login');
-var registerRouter = require('./routes/register');
-
+var accountRouter = require('./routes/account'); 
 
 // setup database connection
 const mongoose = require('mongoose');
@@ -45,8 +43,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 // setup routes
 app.use('/api', apiRouter);
-app.use('/login', loginRouter);
-app.use('/register', registerRouter);
+app.use('/account', accountRouter);
 app.use('/', auth.verifyJWTCookie, indexRouter);
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -54,7 +51,7 @@ app.use(function(req, res, next) {
 });
 
 // error handler
-app.use(function(err, req, res, next) {
+app.use(function(err, req, res) {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get('env') === 'development' ? err : {};
