+++ views/index.mustache
@@ -6,5 +6,6 @@
   <body>
     <h1>{{ title }}</h1>
     <p>Welcome to {{ title }}</p>
+    <div id="logout">Log Out</div>
   </body>
 </html>
