+++ views/login.mustache
@@ -3,7 +3,7 @@
         <title>Login</title>
     </head>
     <body>
-        <form action="/login" method="post">
+        <form action="/account/login" method="post">
             <label for="username">Username</label><br>
             <input type="text" id="username" name="username"><br>
             <label for="password">Password</label><br>
